# docs/api/openapi.locked.yaml
openapi: 3.1.0
info:
  title: FaultMaven API
  description: AI-powered troubleshooting assistant for Engineers, SREs, and DevOps
    professionals
  version: 1.0.0
paths:
...
  /api/v1/knowledge/jobs/{job_id}:
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Get Job Status
      description: "Get the status of a knowledge base ingestion job\n\nArgs:\n  \
        \  job_id: Job identifier\n\nReturns:\n    Job status information"
      operationId: get_job_status_api_v1_knowledge_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Job Status Api V1 Knowledge Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/kb/jobs/{job_id}:
    get:
      tags:
      - knowledge_base
      - knowledge_base
      summary: Get Job Status Kb
      description: Get job status (kb prefix)
      operationId: get_job_status_kb_api_v1_kb_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Job Status Kb Api V1 Kb Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - job_management
      - job_management
      summary: Get Job Status
      description: 'Get job status with proper polling semantics


        Implements consistent job polling with appropriate headers:

        - 200 OK for running/pending jobs with Retry-After header

        - 303 See Other redirect for completed jobs with results

        - 200 OK for failed/cancelled jobs (terminal states)'
      operationId: get_job_status_api_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - job_management
      - job_management
      summary: Cancel Job
      description: 'Cancel a running job


        Attempts to cancel a job if it''s still in a cancellable state.

        Returns 204 No Content on success.'
      operationId: cancel_job_api_v1_jobs__job_id__delete
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs:
    get:
      tags:
      - job_management
      - job_management
      summary: List Jobs
      description: 'List jobs with optional filtering and pagination


        Returns a paginated list of jobs with proper pagination headers.

        Supports filtering by job status.'
      operationId: list_jobs_api_v1_jobs_get
      parameters:
      - name: status_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by job status
          title: Status Filter
        description: Filter by job status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of results
          default: 50
          title: Limit
        description: Maximum number of results
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Result offset for pagination
          default: 0
          title: Offset
        description: Result offset for pagination
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobStatus'
                title: Response List Jobs Api V1 Jobs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobStatus:
      properties:
        job_id:
          type: string
          title: Job Id
        status:
          type: string
          enum:
          - pending
          - running
          - completed
          - failed
          title: Status
        progress:
          anyOf:
          - type: integer
          - type: 'null'
          title: Progress
        result:
          anyOf:
          - type: object
          - type: 'null'
          title: Result
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
      type: object
      required:
      - job_id
      - status
      - created_at
      - updated_at
      title: JobStatus
      description: Async job status tracking model.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...